Add JSONTransformers for NSURL's

+++ VRBCards/VRBCards/VRBConstants.h
@@ -10,6 +10,6 @@
 
 static const NSString *CLASS_PREFIX = @"VRB";
 
-static const CGFloat NAVBAR_HEIGHT = 64.0f;
+static const float NAVBAR_HEIGHT = 64.0f;
 
-static const CGFloat STANDARD_MARGIN = 10.0f;
+static const float STANDARD_MARGIN = 10.0f;
